fix database bug

+++ app/database/models.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, Date, Boolean, ForeignKey, Enum
+from sqlalchemy import Column, Integer, String, Date, Boolean, ForeignKey, Enum, BigInteger
 from sqlalchemy.orm import relationship
 
 import enum
@@ -10,7 +10,7 @@ class User(Base):
     __tablename__ = 'users'
 
     id = Column(Integer, primary_key=True, autoincrement=True)
-    telegram_id = Column(Integer, nullable=False, unique=True)
+    telegram_id = Column(BigInteger, nullable=False, unique=True)
     name = Column(String, nullable=False)
     user_info = Column(String, nullable=True)
